# docs/api/openapi_spec.yml
openapi: 3.0.3
info:
  title: 🧬 Consciousness Web Builder API
  description: |
    Production-ready API for consciousness-engineered AI development platform.

    **Core Capabilities:**
    - 🤖 AI-powered code generation with consciousness patterns
    - 👤 Random user onboarding with intelligence profiling (70-95 IQ)
    - 🚀 Multi-tenant agent orchestration (7 agent types)
    - 🛠️ Dynamic tool fabrication using cognitive kernel
    - 🧠 Meta-cognitive program synthesis with Northstar integration
    - 📊 Real-time metrics and monitoring (1253 users, 1142 projects)
    - 🔄 Autonomous background processing and consciousness loops

    **Architecture:** FastAPI + PostgreSQL + Redis + Celery
    **Security:** JWT authentication with RBAC
    **Monitoring:** Prometheus metrics + structured logging
  version: 1.0.0
  contact:
    name: Consciousness Web Builder Team
    email: api@consciousness-web-builder.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.consciousness-web-builder.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/projects/generate:
    post:
      summary: Generate AI Project with Consciousness Engine
      description: |
        Create AI-powered project using external consciousness engine.

        **Core Consciousness Feature:** Preserved from original system with enhanced async processing.
        - Project Types: web_app, api, agent
        - AI Generation: Consciousness-engine powered
        - Intelligence Context: User intelligence affects generation quality
      tags: [Projects]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
              properties:
                user_id:
                  type: string
                  description: ID of the user creating the project
                project_type:
                  type: string
                  enum: [web_app, api, agent]
                  default: web_app
                  description: Type of project to generate
                requirements:
                  type: string
                  description: Specific requirements for the project
                async_processing:
                  type: boolean
                  default: false
                  description: Process asynchronously for large projects
      responses:
        '201':
          description: Project generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    type: object
                    properties:
                      id:
                        type: string
                      user_id:
                        type: string
                      name:
                        type: string
                        example: "React Web Application"
                      type:
                        type: string
                        example: "web_app"
                      code:
                        type: string
                        description: Generated AI code
                      status:
                        type: string
                        example: "generated"
                      created_at:
                        type: string
                        format: date-time
                      code_length:
                        type: integer
                        description: Length of generated code
                  message:
                        type: string
                        example: "AI-generated project created successfully"
        '202':
          description: Project generation started asynchronously
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Project generation started asynchronously"
                  task_id:
                    type: string
                    example: "task_user123_web_app_123456"
  /api/v1/projects:
    get:
      summary: List AI-Generated Projects
      description: |
        Paginated list of all consciousness-engine-generated projects.

        **Enhanced Features:**
        - Advanced filtering by type, status, user
        - Search functionality
        - Consciousness generation metadata
      tags: [Projects]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: user_id
          in: query
          schema:
            type: string
          description: Filter by user ID
        - name: project_type
          in: query
          schema:
            type: string
            enum: [web_app, api, agent]
          description: Filter by project type
        - name: status
          in: query
          schema:
            type: string
            enum: [generated, completed, failed]
          description: Filter by project status
        - name: search
          in: query
          schema:
            type: string
          description: Search in project names
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        user_id:
                          type: string
                        user_name:
                          type: string
                        name:
                          type: string
                        type:
                          type: string
                        status:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                        code_length:
                          type: integer
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
                  filters:
                    type: object
                    properties:
                      user_id:
                        type: string
                        nullable: true
                      project_type:
                        type: string
                        nullable: true
                      status:
                        type: string
                        nullable: true
  /api/v1/projects/{project_id}:
    get:
      summary: Get Detailed Project Information
      description: Retrieve complete project details including AI-generated code
      tags: [Projects]
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
          description: Unique project identifier
      responses:
        '200':
          description: Project details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  user_id:
                    type: string
                  user_name:
                    type: string
                  name:
                    type: string
                  type:
                    type: string
                  code:
                    type: string
                    description: Full AI-generated code
                  status:
                    type: string
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
                    nullable: true
                  code_length:
                    type: integer
                  code_lines:
                    type: integer
        '404':
          description: Project not found

    delete:
      summary: Delete Project
      description: Remove a project and associated data
      tags: [Projects]
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project deleted successfully
        '404':
          description: Project not found
  /api/v1/projects/{project_id}/regenerate:
    post:
      summary: Regenerate Project with New Requirements
      description: Regenerate existing project using consciousness engine with updated requirements
      tags: [Projects]
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                requirements:
                  type: string
                  description: New requirements for regeneration
      responses:
        '200':
          description: Project regenerated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    type: object
                    description: Regenerated project data
                  message:
                    type: string
                    example: "Project regenerated successfully"
  /api/v1/projects/statistics/overview:
    get:
      summary: Project Generation Analytics
      description: Comprehensive analytics for AI project generation
      tags: [Projects]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Project statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_projects:
                    type: integer
                  type_distribution:
                    type: object
                    additionalProperties:
                      type: integer
                    example: {"web_app": 650, "api": 380, "agent": 112}
                  status_distribution:
                    type: object
                    additionalProperties:
                      type: integer
                    example: {"generated": 890, "completed": 234, "failed": 18}
                  recent_projects:
                    type: integer
                    description: Projects created in last 30 days
                  average_code_lengths:
                    type: object
                    additionalProperties:
                      type: number
                    example: {"web_app": 5420.5, "api": 3890.2, "agent": 7234.8}
                  most_popular_type:
                    type: string
                    example: "web_app"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
...